Generically implement BoxUnboxVmValue with HeapAllocatedData marker trait
